// apps/server/src/infra/files-storage-rest-client/generated/models/file-record-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software File Storage API
 * This is the API documentation for the Schulcloud-Verbund-Software File Storage API
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FileRecordParentType } from './file-record-parent-type';
// May contain unused imports in some cases
// @ts-ignore
import type { FileRecordScanStatus } from './file-record-scan-status';
// May contain unused imports in some cases
// @ts-ignore
import type { PreviewStatus } from './preview-status';

/**
 * 
 * @export
 * @interface FileRecordResponse
 */
export interface FileRecordResponse {
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'parentId': string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'url': string;
    /**
     * 
     * @type {FileRecordScanStatus}
     * @memberof FileRecordResponse
     */
    'securityCheckStatus': FileRecordScanStatus;
    /**
     * 
     * @type {number}
     * @memberof FileRecordResponse
     */
    'size': number;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'creatorId': string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'mimeType': string;
    /**
     * 
     * @type {FileRecordParentType}
     * @memberof FileRecordResponse
     */
    'parentType': FileRecordParentType;
    /**
     * 
     * @type {boolean}
     * @memberof FileRecordResponse
     */
    'isUploading'?: boolean;
    /**
     * 
     * @type {PreviewStatus}
     * @memberof FileRecordResponse
     */
    'previewStatus': PreviewStatus;
    /**
     * 
     * @type {boolean}
     * @memberof FileRecordResponse
     */
    'isCollaboraEditable': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof FileRecordResponse
     */
    'exceedsCollaboraEditableFileSize': boolean;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'deletedSince'?: string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'contentLastModifiedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof FileRecordResponse
     */
    'expiresAt'?: string;
}